requirements.txt
absl-py==2.4.0
asgiref==3.11.1
astunparse==1.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.5
dj-database-url==3.1.2
dj-rest-auth==7.1.1
Django==5.1
django-allauth==65.14.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.3
Faker==40.8.0
flatbuffers==25.12.19
gast==0.7.0
google-auth==2.48.0
google-auth-oauthlib==1.3.0
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.1.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
keras==3.13.2
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.2
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
protobuf==7.34.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
whitenoise==6.12.0
wrapt==2.1.1
